+++ src/gui/guiFormSpecMenu.h
@@ -291,6 +291,9 @@ public:
 
 	~GUIFormSpecMenu();
 
+	static void create(GUIFormSpecMenu **cur_formspec, Client *client,
+		JoystickController *joystick, IFormSource *fs_src, TextDest *txt_dest);
+
 	void setFormSpec(const std::string &formspec_string,
 			const InventoryLocation &current_inventory_location)
 	{
